% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynamicBmaNormalCpp_main.R
\docType{package}
\name{DynamicBmaNormalCpp-package}
\alias{DynamicBmaNormalCpp}
\alias{DynamicBmaNormalCpp-package}
\title{Dynamic Bayesian Model Averaging with C++}
\description{
DynamicBmaNormalCpp implements the Dynamic Bayesian Model technique
of Rafterey et al. (2005) for the normal distribution using C++
}

