% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{em}
\alias{em}
\title{The em algorithm for dynamic bma using the normal distribuiton}
\usage{
em(mean_k, y, squared_error, priors, weights, sigma, maxiter = 1000L)
}
\arguments{
\item{mean_k}{(NumericMatrix) the bias corrected mean of the forecast
for each member of the ensemble (columns) for each
time period (rows)}

\item{y}{(NumericVector) the quantity to be forecast}

\item{squared_error}{(arma::mat) A matrix with the squared errors for each
member of the ensemble (columns) and each time period (rows)}

\item{priors}{(NumericVector) the prior distribution for the weights (chosen
by the user)}

\item{weights}{(NumericVector) the initial weights to use in the algorithm}

\item{sigma}{(double) the intial standard deviation for the normal distribution}

\item{maxiter}{(int) the maximum number of iterations. Defaults to 1000}
}
\value{
A \code{List} with the following elements: \code{weights}, \code{sigma}, and
the \code{log_lik}
}
\description{
The E-M algorithm for Dynamic BMA
}
\details{
For the E-M Algorithm for the Dynamic BMA of Rafterey et al.
(2005) using the normal distribution. This function will conduct
the EM algorithm for maximaziation of the of the
Likelihood function as on p. 1159 of Raftery et al. (2005)
}
\author{
Chandler Lutz
}

